refactor(net): move rule types, parsing, and resolution into network/rules
